put message processor in it's own file
